+++ limbus/app/ViewClass.py
@@ -1,0 +1,7 @@
+from flask import jsonify, wrappers
+
+
+class ViewClass():
+
+    def to_json(self) -> wrappers.Response:
+        return jsonify(self.get_attributes())
