+++ limbus/app/sample/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 sample = Blueprint("sample", __name__)
 
-from .views import *
+from .routes import *
