+++ limbus/app/sample/routes/sample.py
@@ -1,0 +1,19 @@
+from .. import sample
+from flask import render_template
+from flask_login import login_required
+
+from ..views.sample import SampleView
+
+@sample.route("abc/")
+def abc():
+    return "abc"
+
+@sample.route("view/LIMBSMP-<sample_id>")
+@login_required
+def view(sample_id: int):
+    sample = SampleView(sample_id).get_attributes()
+
+    return render_template(
+        "sample/sample/view.html",
+        sample=sample
+    )
