+++ limbus/app/templates/sample/sample/view.html
@@ -7,96 +7,9 @@
 
     <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pb-2 mb-3 border-bottom">
             <h2><i class="fas fa-vial"></i> LIMBSMP-{{ sample.id }}</h2>
-            <div class="btn-toolbar mb-2 mb-md-0">
-                {% if (sample.sample_status.value == "Available") %}
-                    <div class="btn btn-success"><i class="fa fa-check"></i> {{ sample.sample_status.value }}</div>
-                    {% elif (sample.sample_status.value == "Missing") %}
-                    <div class="btn btn-danger"><i class="fa fa-question"></i> {{ sample.sample_status.value }}</div>
-                {% endif %}
-            </div>
+
         </div>
 
-        <div class="btn-toolbar form-group" role="toolbar" aria-label="Toolbar with button groups">
-          <div class="btn-group mr-2" role="group" aria-label="First group">
-            <button type="button" class="btn btn-outline-dark"><i class="fa fa-edit"></i> Edit</button>
-          </div>
-          <div class="btn-group mr-2" role="group" aria-label="Second group">
-            <button type="button" class="btn btn-outline-dark"><i class="fa fa-code-branch"></i> Derive</button>
-          </div>
-          <div class="btn-group mr-2" role="group" aria-label="Third group">
-              <a href="{{ url_for("sample.aliquot", sample_id=sample.id) }}">
-                <button type="button" class="btn btn-outline-dark"><i class="fa fa-code-branch"></i> Aliquot</button>
-              </a>
-          </div>
-            <div class="btn-group" role="group" aria-label="Third group">
-            <button type="button" class="btn btn-outline-dark"><i class="fa fa-print"></i> Print</button>
-          </div>
-        </div>
-
-        <div class="row">
-            <div class="col form">
-                {{ show_value(sample.sample_type, "Sample Type", "text") }}
-                {{ show_value(sample_type.sample_type, "Source Type", "text") }}
-                {{ show_value(sample.disposal_date, "Disposal Instructions", "datetime") }}
-                {{ show_value(sample.disposal_instruction, "Disposal Instructions", "text") }}
-                {{ show_value(protocol.name, "Protocol Name", "text")  }}
-
-            </div>
-            <div class="col">
-                {{ show_value(sample.quantity, "Quantity Available", "text") }}
-
-                {% for term, attr in text_attr %}
-                    {{ show_value(attr.value, term.term, "text")  }}
-                {% endfor %}
-
-                {% for term, _, choice in option_attr %}
-                    {{ show_value(choice.term, term.term,  "text")  }}
-                {% endfor %}
-
-                {{ show_value("Not Stored", "Storage", "text")  }}
-
-                {{ show_value(consent_template.name, "Consent Template", "text")  }}
-                {{ show_value(sample.creation_date, "Creation Date", "datetime") }}
-
-            </div>
-        </div>
-
-
-        <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pb-2 mb-3 border-bottom">
-            <h3><i class="fas fa-file"></i> Associated Documents</h3>
-            <a href="{{ url_for('sample.associate_document', sample_id=sample.id) }}">
-            <div class="btn btn-primary">
-                <i class="fa fa-link"></i> Associate Document</div></a>
-        </div>
-
-                <table id="samples" class="table table-striped table-bordered" style="width:100%">
-            <thead>
-            <tr>
-                <th>Identifier</th>
-                <th>Document Name</th>
-                <th>Document Type</th>
-                <th>Upload Date</th>
-                <th>Actions</th>
-            </tr>
-            </thead>
-            <tbody>
-            {% for _, document in associated_document %}
-                <tr>
-                    <td>LIMBDOC-{{ document.id }}</td>
-                    <td>{{ document.name }}</td>
-                    <td>{{ document.type.value }}</td>
-                    <td>{{ document.upload_date.strftime('%Y-%m-%d') }}</td>
-                    <td>
-                        <a href="{{ url_for('document.view', doc_id=document.id) }}">
-                            <div class="btn btn-primary btn-sm">
-                                <i class="fas fa-eye"></i> View
-                                </div>
-                        </a>
-                    </td>
-                </tr>
-            {% endfor %}
-            </tbody>
-        </table>
